Refactoring option parsing to make testing possible
Put CLI option parsing into a method so that it does not run every
time the script is included.  This will allow us to set settings
manually for testing purposes and stop the script from exiting when
it is used without command line options unless these are explicitly
required by the calling script.