feat: Update feedback for filtering files (#52)
Updates the blocks.sh to support recently added child projects and also
updates the filtering aspect of the script to output more details about
files that don't need to be updated. Allows the XSLT to sort the elements
to be put into the same directory as the transform.xsl file.